Move test to library
