+++ crates/subspace-farmer/src/dsn/tests.rs
@@ -1,10 +1,25 @@
 use super::{sync, DSNSync, NoSync, PieceIndexHashNumber, SyncOptions};
+use crate::bench_rpc_client::{BenchRpcClient, BENCH_FARMER_METADATA};
+use crate::multi_farming::{MultiFarming, Options as MultiFarmingOptions};
+use crate::single_plot_farm::SyncResult;
+use crate::{ObjectMappings, Plot};
+use futures::stream::FuturesUnordered;
+use futures::{SinkExt, StreamExt};
+use num_traits::{WrappingAdd, WrappingSub};
 use rand::Rng;
 use std::collections::BTreeMap;
+use std::net::{Ipv4Addr, SocketAddrV4, TcpListener};
 use std::ops::Range;
 use std::sync::{Arc, Mutex};
-use subspace_core_primitives::{Piece, PieceIndex, PieceIndexHash, PIECE_SIZE};
-use subspace_networking::PiecesToPlot;
+use subspace_archiving::archiver::ArchivedSegment;
+use subspace_core_primitives::{
+    ArchivedBlockProgress, FlatPieces, LastArchivedBlock, Piece, PieceIndex, PieceIndexHash,
+    RootBlock, Sha256Hash, PIECE_SIZE, U256,
+};
+use subspace_networking::libp2p::multiaddr::Protocol;
+use subspace_networking::libp2p::{identity, Multiaddr, PeerId};
+use subspace_networking::{NodeRunner, PiecesToPlot};
+use tempfile::TempDir;
 
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 struct TestDSN(BTreeMap<PieceIndexHash, (Piece, PieceIndex)>);
@@ -121,4 +136,209 @@ async fn no_sync_test() {
     .unwrap();
 
     assert!(result.lock().unwrap().is_empty())
+}
+
+#[tokio::test]
+async fn test_dsn_sync() {
+    let seeder_max_plot_size = 20 * 1024 * 1024 / PIECE_SIZE as u64; // 20M
+    let syncer_max_plot_size = 2 * 1024 * 1024 / PIECE_SIZE as u64; // 20M
+    let request_pieces_size = 20;
+
+    let seeder_base_directory = TempDir::new().unwrap();
+
+    let free_port = (10_000..)
+        .find(|port| TcpListener::bind(SocketAddrV4::new(Ipv4Addr::LOCALHOST, *port)).is_ok())
+        .unwrap();
+    let mut seeder_multiaddr = format!("/ip4/127.0.0.1/tcp/{free_port}")
+        .parse::<Multiaddr>()
+        .unwrap();
+
+    let (mut seeder_archived_segments_sender, seeder_archived_segments_receiver) =
+        futures::channel::mpsc::channel(10);
+    let seeder_client =
+        BenchRpcClient::new(BENCH_FARMER_METADATA, seeder_archived_segments_receiver);
+
+    let object_mappings = tokio::task::spawn_blocking({
+        let path = seeder_base_directory.as_ref().join("object-mappings");
+
+        move || ObjectMappings::open_or_create(path)
+    })
+    .await
+    .unwrap()
+    .unwrap();
+
+    let base_path = seeder_base_directory.as_ref().to_owned();
+    let plot_factory = move |plot_index, public_key, max_piece_count| {
+        let base_path = base_path.join(format!("plot{plot_index}"));
+        Plot::open_or_create(base_path, public_key, max_piece_count)
+    };
+
+    let mut seeder_multi_farming = MultiFarming::new(
+        MultiFarmingOptions {
+            base_directory: seeder_base_directory.as_ref().to_owned(),
+            archiving_client: seeder_client.clone(),
+            farming_client: seeder_client.clone(),
+            object_mappings: object_mappings.clone(),
+            reward_address: subspace_core_primitives::PublicKey::default(),
+            bootstrap_nodes: vec![],
+            listen_on: vec![seeder_multiaddr.clone()],
+            dsn_sync: false,
+        },
+        seeder_max_plot_size * PIECE_SIZE as u64,
+        seeder_max_plot_size,
+        plot_factory,
+        false,
+    )
+    .await
+    .unwrap();
+
+    {
+        let mut last_archived_block = LastArchivedBlock {
+            number: 0,
+            archived_progress: ArchivedBlockProgress::Partial(0),
+        };
+
+        for segment_index in 0..seeder_max_plot_size / 256 {
+            let archived_segment = {
+                let root_block = RootBlock::V0 {
+                    segment_index,
+                    records_root: Sha256Hash::default(),
+                    prev_root_block_hash: Sha256Hash::default(),
+                    last_archived_block,
+                };
+
+                let mut pieces = FlatPieces::new(256);
+                rand::thread_rng().fill(pieces.as_mut());
+
+                ArchivedSegment {
+                    root_block,
+                    pieces,
+                    object_mapping: vec![],
+                }
+            };
+
+            if seeder_archived_segments_sender
+                .send(archived_segment)
+                .await
+                .is_err()
+            {
+                break;
+            }
+            last_archived_block.set_complete();
+        }
+    }
+    seeder_multiaddr.push({
+        schnorrkel::PublicKey::from_bytes(
+            &*seeder_multi_farming.single_plot_farms[0].plot.public_key(),
+        )
+        .map(identity::sr25519::PublicKey::from)
+        .map(identity::PublicKey::Sr25519)
+        .map(PeerId::from)
+        .map(Into::into)
+        .map(Protocol::P2p)
+        .unwrap()
+    });
+
+    let node_runners = futures::future::join_all(
+        std::mem::take(&mut seeder_multi_farming.networking_node_runners)
+            .into_iter()
+            .map(NodeRunner::run),
+    );
+    tokio::spawn(async move {
+        node_runners.await;
+    });
+
+    for _ in 0..10 {
+        let syncer_base_directory = TempDir::new().unwrap();
+        let (_sender, syncer_archived_segments_receiver) = futures::channel::mpsc::channel(10);
+        let syncer_client =
+            BenchRpcClient::new(BENCH_FARMER_METADATA, syncer_archived_segments_receiver);
+
+        let object_mappings = tokio::task::spawn_blocking({
+            let path = syncer_base_directory.as_ref().join("object-mappings");
+
+            move || ObjectMappings::open_or_create(path)
+        })
+        .await
+        .unwrap()
+        .unwrap();
+
+        let base_path = syncer_base_directory.as_ref().to_owned();
+        let plot_factory = move |plot_index, public_key, max_piece_count| {
+            let base_path = base_path.join(format!("plot{plot_index}"));
+            Plot::open_or_create(base_path, public_key, max_piece_count)
+        };
+
+        let mut syncer_multi_farming = MultiFarming::new(
+            MultiFarmingOptions {
+                base_directory: syncer_base_directory.as_ref().to_owned(),
+                archiving_client: syncer_client.clone(),
+                farming_client: syncer_client.clone(),
+                object_mappings: object_mappings.clone(),
+                reward_address: subspace_core_primitives::PublicKey::default(),
+                bootstrap_nodes: vec![seeder_multiaddr.clone()],
+                listen_on: vec!["/ip4/127.0.0.1/tcp/0".parse().unwrap()],
+                dsn_sync: false,
+            },
+            syncer_max_plot_size * PIECE_SIZE as u64,
+            syncer_max_plot_size,
+            plot_factory,
+            false,
+        )
+        .await
+        .unwrap();
+
+        let node_runners = futures::future::join_all(
+            std::mem::take(&mut syncer_multi_farming.networking_node_runners)
+                .into_iter()
+                .map(NodeRunner::run),
+        );
+        tokio::spawn(async move {
+            node_runners.await;
+        });
+
+        let expected_total_pieces = syncer_max_plot_size;
+        let range_size = PieceIndexHashNumber::MAX / expected_total_pieces * request_pieces_size;
+        let mut futures = syncer_multi_farming
+            .single_plot_farms
+            .iter()
+            .map(|farming| {
+                farming.dsn_sync(syncer_max_plot_size, expected_total_pieces, range_size)
+            })
+            .collect::<FuturesUnordered<_>>();
+
+        while let Some(result) = futures.next().await {
+            let SyncResult {
+                expected_range:
+                    Range {
+                        start: expected_start,
+                        end: expected_end,
+                    },
+                got_range: Range { start, end },
+                total_pieces,
+                public_key,
+            } = result.unwrap();
+            let shift_to_middle =
+                |n: U256, pub_key| n.wrapping_sub(pub_key).wrapping_add(&U256::MIDDLE);
+
+            let public_key = U256::from_big_endian(&public_key);
+            let expected_start = shift_to_middle(expected_start, &public_key);
+            let expected_end = shift_to_middle(expected_end, &public_key);
+            let start = shift_to_middle(start, &public_key);
+            let end = shift_to_middle(end, &public_key);
+
+            assert!(expected_start <= start && start <= expected_end);
+            assert!(expected_start <= end && end <= expected_end);
+            assert!(
+                expected_total_pieces / 10 * 9 < total_pieces
+                    && total_pieces < expected_total_pieces / 10 * 11
+            );
+        }
+
+        syncer_client.stop().await;
+    }
+
+    drop(seeder_archived_segments_sender);
+    seeder_client.stop().await;
+    seeder_multi_farming.wait().await.unwrap();
 }
