+++ crates/subspace-farmer/src/lib.rs
@@ -22,6 +22,7 @@
 //! 64-bit unsigned integers.
 
 pub(crate) mod archiving;
+pub mod bench_rpc_client;
 pub(crate) mod commitments;
 pub(crate) mod dsn;
 pub(crate) mod farming;
@@ -39,7 +40,6 @@ pub mod ws_rpc_server;
 
 pub use archiving::{Archiving, ArchivingError};
 pub use commitments::{CommitmentError, Commitments};
-pub use dsn::PieceIndexHashNumber;
 pub use farming::{Farming, FarmingError};
 pub use identity::Identity;
 pub use jsonrpsee;
@@ -48,4 +48,3 @@ pub use object_mappings::{ObjectMappingError, ObjectMappings};
 pub use plot::{retrieve_piece_from_plots, PieceOffset, Plot, PlotError, PlotFile};
 pub use plotting::plot_pieces;
 pub use rpc_client::{Error as RpcClientError, RpcClient};
-pub use single_plot_farm::SyncResult;
